DB update

+++ BmeModels/Role.cs
@@ -4,8 +4,7 @@ namespace BmeModels
     public class Role
     {
         public Role(){}
-
-        public sbyte Id { get; set; }
+        public int Id { get; set; }
         public string Title { get; set; } = null!;
         public string? Description { get; set; }
     }
